% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conspire.R
\name{conspire}
\alias{conspire}
\title{Generate a pseudo-random conspiracy theory}
\usage{
conspire()
}
\value{
a character string containing a conspiracy theory
}
\description{
Generate a pseudo-random conspiracy theory
}
\examples{
conspirator::conspire()
}
